[FINISH] firecamp !!

+++ project/game_engine/entity/object.h
@@ -36,8 +36,8 @@ public:
     explicit Object(const char *path);
     Object(float size, float height);
 
-    static std::shared_ptr<Object> make(const char *path, const std::shared_ptr<Renderer>& renderer = nullptr);
-    static std::shared_ptr<Object> make(float size, float height, const std::shared_ptr<Renderer>& renderer = nullptr);
+    static std::shared_ptr<Object> make(const char *path, const std::shared_ptr<Renderer> &renderer = nullptr);
+    static std::shared_ptr<Object> make(float size, float height, const std::shared_ptr<Renderer> &renderer = nullptr);
 
     // this weird line is there so that the overloads of Entity::setPosition stay despite the override
     using Entity::setPosition;
@@ -45,9 +45,9 @@ public:
 
     void setColor(const glm::vec3 &color);
 
-    void setScale(int scale);
+    void setScale(float scale);
 
-    void setOctreeNode(const std::shared_ptr<Octree>& node);
+    void setOctreeNode(const std::shared_ptr<Octree> &node);
 
     [[nodiscard]] float getHeight() const;
